NUMBERS-186 updated logic in replaceAll
